+++ src/lib/components/DropDown/DropDown.js
@@ -82,14 +82,20 @@ const DropDown = React.forwardRef((props, ref) => {
       function keyboardControls(event) {
          const active = document.activeElement
          const target = event.target
+         const {index, length} = getActiveIndex(active, elements)
+         const {shiftKey, key} = event
 
          // Exit if elements are not focusable.
          if (!containerRef.current.contains(active) || elements.length === 0) {
+            switch (key) {
+               // Exit if esc and no focusable elements.
+               case 'Esc':
+               case 'Escape': // Escape
+                  setExpanded(false)
+                  break
+            }
             return
          }
-
-         const {index, length} = getActiveIndex(active, elements)
-         const {shiftKey, key} = event
 
          if (shiftKey) {
             // Shift + Tab
@@ -154,7 +160,7 @@ const DropDown = React.forwardRef((props, ref) => {
          document.addEventListener('click', clickOutside)
          document.addEventListener('keyup', focusOutside)
          document.addEventListener('keydown', keyboardControls)
-         containerRef.current.addEventListener('focusout', focusOut)
+         //containerRef.current.addEventListener('focusout', focusOut)
          loaded.current = true
       }
 
@@ -162,56 +168,9 @@ const DropDown = React.forwardRef((props, ref) => {
          document.removeEventListener('click', clickOutside)
          document.removeEventListener('keyup', focusOutside)
          document.removeEventListener('keydown', keyboardControls)
-         containerRef.current.removeEventListener('focusout', focusOut)
+         //containerRef.current.removeEventListener('focusout', focusOut)
       }
    }, [])
-
-   /**
-    * Search the menu elements by first letter.
-    *
-    * @param {string} current The currently active element.
-    * @param {string} char    The character to search.
-    * @returns null
-    */
-   function searchByFirstLetter(current, char) {
-      const elements = menuRef.current.querySelectorAll(focusable)
-      let start = 0
-      let index = 0
-
-      if (char.length > 1 || !elements) {
-         return
-      }
-
-      const array = Array.prototype.slice.call(elements)
-
-      // First letters.
-      const letters =
-         array &&
-         array.map((item) => {
-            return item?.textContent
-               ? item.textContent.trim()[0].toLowerCase()
-               : ''
-         })
-
-      // Get start item from the position of the current item.
-      start = array.indexOf(current) + 1
-      if (start >= array.length) {
-         start = 0
-      }
-
-      // Check menu elements.
-      index = letters.indexOf(char.toLowerCase(), start)
-
-      // Search from beginning.
-      if (index === -1) {
-         index = letters.indexOf(char.toLowerCase(), 0)
-      }
-
-      // Match found, set focus
-      if (index > -1) {
-         setFocus(array[index])
-      }
-   }
 
    /**
     * Click handler to toggle the dropdown menu.
@@ -219,7 +178,7 @@ const DropDown = React.forwardRef((props, ref) => {
     * @param {Event} event The click event.
     */
    function toggleMenu(e) {
-      e.currentTarget.blur()
+      //e.currentTarget.blur()
       const elements = menuRef.current.querySelectorAll(focusable)
       if (elements && !expanded) {
          setFocus(elements[0])
@@ -279,6 +238,53 @@ const DropDown = React.forwardRef((props, ref) => {
    }
 
    /**
+    * Search the menu elements by first letter.
+    *
+    * @param {string} current The currently active element.
+    * @param {string} char    The character to search.
+    * @returns null
+    */
+   function searchByFirstLetter(current, char) {
+      const elements = menuRef.current.querySelectorAll(focusable)
+      let start = 0
+      let index = 0
+
+      if (char.length > 1 || !elements) {
+         return
+      }
+
+      const array = Array.prototype.slice.call(elements)
+
+      // First letters.
+      const letters =
+         array &&
+         array.map((item) => {
+            return item?.textContent
+               ? item.textContent.trim()[0].toLowerCase()
+               : ''
+         })
+
+      // Get start item from the position of the current item.
+      start = array.indexOf(current) + 1
+      if (start >= array.length) {
+         start = 0
+      }
+
+      // Check menu elements.
+      index = letters.indexOf(char.toLowerCase(), start)
+
+      // Search from beginning.
+      if (index === -1) {
+         index = letters.indexOf(char.toLowerCase(), 0)
+      }
+
+      // Match found, set focus
+      if (index > -1) {
+         setFocus(array[index])
+      }
+   }
+
+   /**
     * Get the current index position for the active element.
     *
     * @param   {HTMLElement} el       The current element to compare.
@@ -315,7 +321,7 @@ const DropDown = React.forwardRef((props, ref) => {
     */
    function createMarkup(html) {
       return {
-         __html: html
+         __html: `<span>${html}</span>`
       }
    }