+++ src/pages/home/MenuSearch.js
@@ -1,22 +1,22 @@
 export default function MenuSearch() {
    return (
       <ul className="text-sm">
-         <li className="border-b py-1">
+         <li className="border-b border-opacity-50 py-1">
             <a href="#" className="block w-full text-left p-1 hover:underline">
                About Us
             </a>
          </li>
-         <li className="py-1">
+         <li className="border-b border-opacity-50 py-1">
             <a href="#" className="block w-full text-left p-1 hover:underline">
                Board of Directors
             </a>
          </li>
-         <li className="border-b py-1">
+         <li className="border-b border-opacity-50 py-1">
             <a href="#" className="block w-full text-left p-1 hover:underline">
                Our Locations
             </a>
          </li>
-         <li className="border-b py-1">
+         <li className="border-b border-opacity-50 py-1">
             <a href="#" className="block w-full text-left p-1 hover:underline">
                Contact Us
             </a>