+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+**0.5.1**
+
+-  UPDATE: Updated component styling defaults and `config` props.
+-  UPDATE: Updated demo docs.
+-  NEW: Added testing instructions.
+-  FIX: Fixed issue with esc not working for elements without a focusable element.
+-  FIX: Issue with esc not working for elements without a focusable element.
+
 **0.2.2**
 
 UPDATE: Improved default styles and config options.